datastore: iter -> state_view + move in StateView
