+++ crates/core/src/db/datastore/locking_tx_datastore/state_view.rs
@@ -1,17 +1,194 @@
-use std::ops::RangeBounds;
-
-use spacetimedb_primitives::{ColList, TableId};
-use spacetimedb_sats::{AlgebraicValue, ProductValue};
-
-use crate::{db::db_metrics::DB_METRICS, execution_context::ExecutionContext};
-
 use super::{
     btree_index::BTreeIndexRangeIter,
     committed_state::{get_committed_row, CommittedIndexIter, CommittedState},
-    mut_tx::StateView as _,
     tx_state::{RowState, TxState},
     DataRef, RowId,
 };
+use crate::db::datastore::Result;
+use crate::{
+    db::{
+        datastore::system_tables::{
+            StColumnFields, StColumnRow, StConstraintFields, StConstraintRow, StIndexFields, StIndexRow,
+            StSequenceFields, StSequenceRow, StTableFields, StTableRow, SystemTable, ST_COLUMNS_ID, ST_CONSTRAINTS_ID,
+            ST_INDEXES_ID, ST_SEQUENCES_ID, ST_TABLES_ID,
+        },
+        db_metrics::DB_METRICS,
+    },
+    error::TableError,
+    execution_context::ExecutionContext,
+};
+use spacetimedb_lib::Address;
+use spacetimedb_primitives::{ColList, TableId};
+use spacetimedb_sats::{
+    db::def::{ColumnSchema, ConstraintSchema, IndexSchema, SequenceSchema, TableSchema},
+    AlgebraicValue, ProductValue,
+};
+use std::{borrow::Cow, ops::RangeBounds};
+
+// StateView trait, is designed to define the behavior of viewing internal datastore states.
+// Currently, it applies to: CommittedState, MutTxId, and TxId.
+pub trait StateView {
+    fn get_schema(&self, table_id: &TableId) -> Option<&TableSchema>;
+
+    fn table_id_from_name(&self, table_name: &str, database_address: Address) -> Result<Option<TableId>> {
+        self.iter_by_col_eq(
+            &ExecutionContext::internal(database_address),
+            &ST_TABLES_ID,
+            ColList::new(StTableFields::TableName.col_id()),
+            AlgebraicValue::String(table_name.to_owned()),
+        )
+        .map(|mut iter| {
+            iter.next()
+                .map(|row| TableId(*row.view().elements[0].as_u32().unwrap()))
+        })
+    }
+
+    fn iter<'a>(&'a self, ctx: &'a ExecutionContext, table_id: &TableId) -> Result<Iter<'a>>;
+
+    fn table_exists(&self, table_id: &TableId) -> Option<&str>;
+
+    /// Returns an iterator,
+    /// yielding every row in the table identified by `table_id`,
+    /// where the values of `cols` are contained in `range`.
+    fn iter_by_col_range<'a, R: RangeBounds<AlgebraicValue>>(
+        &'a self,
+        ctx: &'a ExecutionContext,
+        table_id: &TableId,
+        cols: ColList,
+        range: R,
+    ) -> Result<IterByColRange<'a, R>>;
+
+    fn iter_by_col_eq<'a>(
+        &'a self,
+        ctx: &'a ExecutionContext,
+        table_id: &TableId,
+        cols: ColList,
+        value: AlgebraicValue,
+    ) -> Result<IterByColEq<'_>> {
+        self.iter_by_col_range(ctx, table_id, cols, value)
+    }
+
+    fn schema_for_table(&self, ctx: &ExecutionContext, table_id: TableId) -> Result<Cow<'_, TableSchema>> {
+        if let Some(schema) = self.get_schema(&table_id) {
+            return Ok(Cow::Borrowed(schema));
+        }
+
+        // Look up the table_name for the table in question.
+        let table_id_col = ColList::new(StTableFields::TableId.col_id());
+
+        // let table_id_col = ColList::new(.col_id());
+        let value: AlgebraicValue = table_id.into();
+        let rows = self
+            .iter_by_col_eq(ctx, &ST_TABLES_ID, table_id_col, table_id.into())?
+            .collect::<Vec<_>>();
+        let row = rows
+            .first()
+            .ok_or_else(|| TableError::IdNotFound(SystemTable::st_table, table_id.into()))?;
+        let el = StTableRow::try_from(row.view())?;
+        let table_name = el.table_name.to_owned();
+        let table_id = el.table_id;
+
+        // Look up the columns for the table in question.
+        let mut columns = self
+            .iter_by_col_eq(
+                ctx,
+                &ST_COLUMNS_ID,
+                ColList::new(StColumnFields::TableId.col_id()),
+                value,
+            )?
+            .map(|row| {
+                let el = StColumnRow::try_from(row.view())?;
+                Ok(ColumnSchema {
+                    table_id: el.table_id,
+                    col_pos: el.col_pos,
+                    col_name: el.col_name.into(),
+                    col_type: el.col_type,
+                })
+            })
+            .collect::<Result<Vec<_>>>()?;
+
+        columns.sort_by_key(|col| col.col_pos);
+
+        // Look up the constraints for the table in question.
+        let mut constraints = Vec::new();
+        for data_ref in self.iter_by_col_eq(
+            ctx,
+            &ST_CONSTRAINTS_ID,
+            ColList::new(StConstraintFields::TableId.col_id()),
+            table_id.into(),
+        )? {
+            let row = data_ref.view();
+
+            let el = StConstraintRow::try_from(row)?;
+            let constraint_schema = ConstraintSchema {
+                constraint_id: el.constraint_id,
+                constraint_name: el.constraint_name.to_string(),
+                constraints: el.constraints,
+                table_id: el.table_id,
+                columns: el.columns,
+            };
+            constraints.push(constraint_schema);
+        }
+
+        // Look up the sequences for the table in question.
+        let mut sequences = Vec::new();
+        for data_ref in self.iter_by_col_eq(
+            ctx,
+            &ST_SEQUENCES_ID,
+            ColList::new(StSequenceFields::TableId.col_id()),
+            AlgebraicValue::U32(table_id.into()),
+        )? {
+            let row = data_ref.view();
+
+            let el = StSequenceRow::try_from(row)?;
+            let sequence_schema = SequenceSchema {
+                sequence_id: el.sequence_id,
+                sequence_name: el.sequence_name.to_string(),
+                table_id: el.table_id,
+                col_pos: el.col_pos,
+                increment: el.increment,
+                start: el.start,
+                min_value: el.min_value,
+                max_value: el.max_value,
+                allocated: el.allocated,
+            };
+            sequences.push(sequence_schema);
+        }
+
+        // Look up the indexes for the table in question.
+        let mut indexes = Vec::new();
+        for data_ref in self.iter_by_col_eq(
+            ctx,
+            &ST_INDEXES_ID,
+            ColList::new(StIndexFields::TableId.col_id()),
+            table_id.into(),
+        )? {
+            let row = data_ref.view();
+
+            let el = StIndexRow::try_from(row)?;
+            let index_schema = IndexSchema {
+                table_id: el.table_id,
+                columns: el.columns,
+                index_name: el.index_name.into(),
+                is_unique: el.is_unique,
+                index_id: el.index_id,
+                index_type: el.index_type,
+            };
+            indexes.push(index_schema);
+        }
+
+        Ok(Cow::Owned(TableSchema::new(
+            table_id,
+            table_name,
+            columns,
+            indexes,
+            constraints,
+            sequences,
+            el.table_type,
+            el.table_access,
+        )))
+    }
+}
 
 pub struct Iter<'a> {
     ctx: &'a ExecutionContext<'a>,
